requirements.txt
selenium~=4.7.2
pytest~=7.2.1
python-dotenv~=0.21.1
Faker~=19.12.0
requests~=2.31.0